Refactor builtin source loading into a Provider + etc.
DataSources can now belong to *any* Provider, as it should be. Compiled
sources are still cached in memory for sharing amongst multiple sheets.
Builtin sources just belong to a specific Provider called :wish.

We still don't really handle errors loading a source, but we at least
get the signal, now, so we *could* handle it....