Refactor fn compilation to work in :advanced mode
First, we manually destructure input, because let and fn are
apparently macros that don't play nicely in :advanced mode.
Our destructuring usage is very straightforward anyway, and
hopefully datasources won't need functions for anything super
complicated anyway—that's why we're using real re-frame stuff for
sheets!

Next, we have to export cljs.core/truth_ as a symbol, even though
it's a macro. I couldn't tell you why it's like this, but we need
to export this because (cond) uses it internally.

We also introduce the use of clojure.walk/postwalk to replace fn
references with the ones we've explicitly exported, so those calls
can work, and, finally, handle `=` in function names properly.

Compiling with the fns replaced produces a ton of warnings, but
because of our export use, the compiled fns do work, so I've suppressed
the warnings.