Fix potential double-load opening directly to a sheet
It was a timing issue, but happened pretty consistently. Basically
what would happen is that opening directly to a sheet would cause
sheets/ensuring-loaded to fetch the sheet immediately, and it would
finish before the initial listing finished. This could cause the
sheet to briefly show if the data-source also loaded fast.

Then, the listing would finish, and :add-sheets would overwrite
the sheet entry, throwing the page back into the loader, and causing
sheets/ensure-loaded to have to load the sheet again.

Simple fix is to not overwrite existing entries in :add-sheets!