Refactor dnd5e source into separate files and add script to combine
Using planck to combine the files in clojurescript is neat, but
we should maybe just use nodejs since it's probably available on
travis without any futsing, has more mature file system tools, etc.
Planck has the File interface, but we can't use the normal Java methods
on it, which Clojure sort of assumes, so it's much harder to work with.