Recursively apply options in order of *able to apply*
Some options rely on other options being applied first for their
associated feature to be available. By dividing options into a set
of *currently applyable* and *not yet applyable* we have a sort of
implicit dependency graph that naturally applies options in the order
they *need* to be applied while maintaining a basically linear time
complexity for (apply-options)!