Begin adding more Monk features + start using :bonus syntax
By marking :bonus actions and :reaction options as attrs, we could
potentially surface them in the UI at some point. I've also introduced
the :consumes property of features, which point to a limited-use that
they consume some amount of. This could also be surfaced in the UI
in a convenient way.