Add + support data source selection UI; prefix sheet ids
Here, "support" data source selection means fixing the handling of
loading multiple sources.

The provider-specific-id part of a sheet/resource ID is now prefixed
with "w" to handle any pro-ids that start with a number, because the
clojure reader for whatever reason doesn't support such keywords
(even though the keyword function is perfectly cool with it, and they
work fine if they don't have a namespace...).

Now the bad news: reading each datasource separately doesn't actually
work correctly, because the :!provide-options directive will never
find the feature into which to install the options. Derp.

It's possible we could preserve the :deferred-options map and
merge it together as part of the (composite) factory function....