Add a virtual-list widget
Refs #7

We don't use it yet, though. I had hoped to use it for many-:values
options, but it's just not compatible with reagent-forms. It's still
a candidate, for sure, but will take some time to refactor.

A stopgap will be to adjust limited-select so we know whether or not
the row is selected when rendering, and not render the :desc if
unselected.