Add a basic custom item creator
Refs #15

Currently supports simple stackables (Ammunition, Potion) and "Other"
for unidentified or generic objects. We should also have basic support
for armor, weapons, and generic "gear."

This commit also de-emphasizes the explicit :stacks? property on items
in favor of :type-based stacking. We will still support it, though,
for sheets that don't want to or can't use :type for some reason and
need the explicit :stack? flag.