Refactor sheet-load error handling to facilitate specific solutions
Refs #25

In particular, we no longer show "Loading" if there was an error,
because we probably aren't. Instead, we delete the :sheet-sources
entry if it had an :err on a successful load, and when initiating
a load.

This commit also establishes the convention that if the sheet error
is an (ex-info), the associated data map is a set of flags indicating
things that might have gone wrong. Currently we support :permissions?
which means that the active user may not have permission to view the
file; the resolution of this is provider-specific---for gdrive, we
may try letting the user open the file with Picker as a way of granting
the app permission....