Support :available? on options; add prerequisite filters for Warlock
This is mentioned in the comments a few times, but reagent-forms
unfortunately doesn't play nicely with dynamically changing option
sets for :list, so we have to use the :visible? fn to evaluate
a subscription that checks for the status of :available?, and instead
of just removing not-:available? options, store the evaluated value.