Refactor sheet-specific style handling
Includes new macros to use in place of the default cljs-css-modules'
defstyle: `defclass` and `defstyled`. Whereas the default `defstyle`
generates multiple classes that you have to pull out by key—in a
somewhat unintuitive way, mind you: if the root style is `[:.foo]`,
you have to use `(:foo styles)`—our `defclass` simply generates a
single class whose name is `def`'d onto the `style-id` you provided.
`defclass` also lets you pass styles directly, which is pretty
intuitive:

```clojure
(defclass big {:font-size "300%"})

[:div {:class big} "Big text!"]
```

`defstyled` is basically just a convenience for when you're generating
a style for a component that probably won't need other attributes.
Basically, it generates `{:class (defclass)}`, so you could always
use `(merge)` or `(assoc)` to add other attributes if you were so
inclined.

I chose this approach over a "styled components"-type approach where
you might use a `[Big "Big Text!"]` syntax because it's a bit more
flexible—you can combine classes in the normal HTML/CSS way with
`(defclass)` and the type of element for `(defstyled)` is controlled
at the use site.

I wrote this instead of switching to Stylefy because I don't like
their syntax and think their decision to call psuedoclasses "modes"
and have a special syntax for it is silly—at the end of the day,
we're writing CSS, so the closer our mental model can be to that,
the better. Artifically distancing ourselves just makes it harder
for new engineers to a project to get into it, in my opinion.