Add `[:action]` attrs for features that use an Action
We would like to surface regular actions in addition to bonus
actions and reactions—the first step is to annotate them!

Note that the Dragonborn breath weapon does not provide an `[:action]`
attr, since it provides the `[:attack]` attr. Right now we're assuming
that all provided `[:attack]`s use an Action.