Refactor wish.style ns to avoid hidden dependency on css-modules
For whatever reason, the transitive :require dependency wasn't picked
up by the compiler as a problem, but did in fact break many things---in
particular, parts of the code like wish.core.init which are supposed
to be ^:export'd were... not, and even in dev mode it complained.

By using `(:require-macros [wish.style])` in the `wish.style` cljs
namespace, we can :refer to the defclass/defstyled macros like we
would in Clojure, and can do the cljs-css-modules.macro once, in
the `wish.style` cljs ns where it makes sense, and not have to worry
about it everywhere else we might want to use the macros. Nice!

Credit for this solution goes to:
https://groups.google.com/forum/#!msg/clojurescript/o0Q8Gs_0HYQ/GFFXOXOVBgAJ