Fixed: removing :primary-only? features when not filtering on entity
I believe that part of the sub was written before we included the
extra keys from :meta/classes in the :classes subscription, which
is why we stupidly had to provide :primary?. By not providing :primary?
we can just ask the :wish/container directly whether it's :primary?,
instead of having to rely on potentially-not-provided, third-party
information.