5e: Support editing custom items (to rename, etc.)
Pretty large refactor, moving that whole (250+ lines!) overlay to
its own ns. But, this is a pretty long-overdue refactor, and there
are probably a few other overlays we should consider moving as well.

Also, in adding the new test ns I refactored the runner to use
doo-all-tests instead of manually specifying each one. I'm not sure
what issues I had with it before that made me avoid it, but it seems
to be working okay—and it's much more maintainable now!