Support offline loading in production mode
Refs #27

We don't recover yet from being offline, but if we fail to load
the initial gapi we go into "gapi unavailable" mode. Saving sheets
logs an error but otherwise seems to not break things. However,
we will need to somehow persist pending changes when we regain
connection (after loading gapi, of course!) or when we come back
on the same client later.