Initial refactor to move sheet querying into the Provider interface
This now can show "Loading" in the sheet browser sooner, since Wish
has more insight into the provider's state. Currently querying will
only be requested when the provider reports that it's ready, however,
and there's no way for the Caching wrapper to intercept that state,
so though the sheet list is cached, the cache is not used in the
offline state.

So, our next refactor will be to make use of init! to move state
management into the Provider interface as well. Basically, the less
the provider directly dispatches events, the easier it will be for
compositing providers such as CachingProvider to override state.