Save cache dirtiness; when cache is dirty, load it directly
Refs #27

Pretty simple solution here. We mark a sheet dirty when we first
dispatch the save event, and clean it on successful write.
If we had previously failed to save a sheet when loading it later,
we'll see that it's still dirty and load from cache directly instead
of trying to load from network. This ensures that we won't accidentally
overwrite a dirty sheet with old data. It doesn't help cross-device,
but it'll have to do for now.

The final step will be to try to resolve any `dirty` flags eagerly
when we regain network.