Add service worker + PWA support (#64)
* Configure figwheel to build a service worker

Refs #27

* Add very rough, basic support for offline mode via service worker!

Refs #27

Also, sheets and data-sources from Drive are cached in local-storage,
since we can't seem to use service-worker cache to load the drive API
js. That's probably fair, and certainly acceptable behavior.

* Remove the compiled worker.js

* Support offline loading in production mode

Refs #27

We don't recover yet from being offline, but if we fail to load
the initial gapi we go into "gapi unavailable" mode. Saving sheets
logs an error but otherwise seems to not break things. However,
we will need to somehow persist pending changes when we regain
connection (after loading gapi, of course!) or when we come back
on the same client later.

* Initial refactor to move sheet querying into the Provider interface

This now can show "Loading" in the sheet browser sooner, since Wish
has more insight into the provider's state. Currently querying will
only be requested when the provider reports that it's ready, however,
and there's no way for the Caching wrapper to intercept that state,
so though the sheet list is cached, the cache is not used in the
offline state.

So, our next refactor will be to make use of init! to move state
management into the Provider interface as well. Basically, the less
the provider directly dispatches events, the easier it will be for
compositing providers such as CachingProvider to override state.

* Change provider/init! to return a channel indicating provider state

:put-provider-state! is now only called by the `providers` ns, so
CachingProvider can override its base Provider's state. This introduces
a new :cache-only state which is pretty self-explanatory.

The new semantics of init! are that it should be idempotent if the
Provider is already initialized, so we've restored the hot-reload
code in `wish.core`.

* Improve robustness of sheet-browser "Loading" state

Finally!

* Move up "create a new sheet" button so it doesn't bounce around

* Use cleaner cond-> syntax

* Watch network state: store in db and init! providers when online

Refs #27

Not sure what to use `:online?` for yet, but we might as well have
it in the DB.

* Try to reload gapi when network returns if it failed previously

Refs #27

* Save cache dirtiness; when cache is dirty, load it directly

Refs #27

Pretty simple solution here. We mark a sheet dirty when we first
dispatch the save event, and clean it on successful write.
If we had previously failed to save a sheet when loading it later,
we'll see that it's still dirty and load from cache directly instead
of trying to load from network. This ensures that we won't accidentally
overwrite a dirty sheet with old data. It doesn't help cross-device,
but it'll have to do for now.

The final step will be to try to resolve any `dirty` flags eagerly
when we regain network.

* Eagerly attempt to persist "dirty" sheets when we regain network

Refs #27

* Prevent dup provider init in debug; make gapi init! more robust

Refs #27

* Add preliminary PWA manifest

* Use (promise-chan) instead of (chan) in gdrive

I wish I'd googled this sooner. This is exactly what I wanted out
of these channels, and simplifies a lot of things