Don't render non-visible instead of animating height; animate headers
The built in animateHeight option is garbage for our purposes; it
sets a fixed height on the element and doesn't react to changes in the
element's height. Instead, we just ereplace pages that aren't selected
(or aren't part of the current swipe) with an empty `[:div]`.

We're currentlty using `scrollIntoView()` to animate the selected header
into position. This has the added side-effect of animating to the top
when swiping between pages. It's not originally intended, but it is
somewhat nicer than popping to the bottom of the new page, which is
what we had before (thanks to the empty div trick above).