5e: Collapse abilities pane into a tab on smartphones; support swiping between tabs (#72)
* 5e: collapse the abilities pane into a tab on smartphones

* Add very rough, initial react-swipeable-views integration

Includes experimentally building the dependency, since both cljsjs
and the react-swipeable-views projects are radio silence on my PRs

* Move react-swipeable-views install to before_install step

* Clone cljsjs-packages via https instead of ssh

Woops

* Install boot, too. Why not.

* Fix boot version reference

* cd to the right place

* Print out boot version

* Remove now-unnecessary boot build step for react-swipeable-views

Naturally, mere hours after I get boot to work on travis, cljsjs
finally publishes my package. Oh well!

* Let `swipeable` connect to a ratom or to re-frame events

* Don't render non-visible instead of animating height; animate headers

The built in animateHeight option is garbage for our purposes; it
sets a fixed height on the element and doesn't react to changes in the
element's height. Instead, we just ereplace pages that aren't selected
(or aren't part of the current swipe) with an empty `[:div]`.

We're currentlty using `scrollIntoView()` to animate the selected header
into position. This has the added side-effect of animating to the top
when swiping between pages. It's not originally intended, but it is
somewhat nicer than popping to the bottom of the new page, which is
what we had before (thanks to the empty div trick above).

* Cleanup, comment, document

* Support swiping between action pages

* Add min-height to action pages to facilitate swiping

* Tweak layout for consistency

Adding "width: 100%" to tablets as well as smartphones prevents
the Features tab from being wider than anything else and triggering
a reflow when we select it.

Using 92% instead of 90% on smartphones seems to prevent clipping
more consistently, although this may just be an artifact of chrome's
debug rendering....