5e: prevent selecting otherwise-selected skill proficiencies
This also includes a change to feature option availability for the
5e builder such that chosen options for a feature are *always*
"available" so that they can be unselected instead of just disappearing
(if their availability is based on being selected).

Though this works well for cases like these where proficiency can be
granted and not just chosen, it might be nice to have a more general
mechanism for things like spells (see #70).