Prevent layout occasionally not expanding to its full max-width
Not sure why this only occasionally repro'd. It was somewhat consistent
on my "Test Tiefling Warlock" sheet, but never saw it on any other.

What seemed to be happening was that the sides were expanding as
intended by their @media styles when >1100px browser width, but the
outer layout, for whatever reason, was not expanding to its 1100px
max-width, so the layout would wrap. Bizarre.

We could also only use `flex/wrap` on smartphone-size displays, but this
seems to be working.