Add :wish/sort to features added by :+features entity mod
Refs #81

Any instanced features added this way (eg: the new :abi-or-feat)
get the `:wish/sort` by compile-entity as normal, but in the process
of apply-entity-mod that `:wish/sort` gets conj'd onto the base
feature entry's `:wish/sorts`, and that gets parceled out when
the instances are inflated in the subscription.

If `:wish/sorts` is guaranteed to be the same as
`(count :wish/instances)` (it probably should be), we could probably
refactor to get rid of `:wish/instances` in favor of just relying on that.