Add `:wish/sort`-ing for instanced :!provide-features
Refs #81

It's a bit tricky, but basically apply-feature-options picks the
right `:wish/sort` for the feature that's "providing" the directives
(which would otherwise have a list of sorts in `:wish/sorts`) based
on its instance id.

Right now we have a very inefficient compare of the sort vectors
so that we can get ([0 0] [0 0 1] [4 0] [4 0 1]) as you would
expect, instead of ([0 0] [4 0] [0 0 1] [4 0 1]). Shouldn't be too
hard to optimize that....