Throttle shell updated notifications
Fixes #86

I *think* this should fix it, by giving time for all assets to
update before issuing the notification.

This commit also moves the cache opening to *before* the request,
which will hopefully reduce the time between receiving the response
and it arriving in the cache (if that was a factor). I tried returning
the response when the cache put finishes via `.then`, but that didn't
seem to work properly (the request never resolved for some reason).