Persist last-update client-side for more robust version checking
Fixes #86

Part of the problem was that the worker was pretty much always sending
the latest version on :ready, because the value was there in the
atom even on refresh. So, obviously, the browser needs to remember which
version it has seen and ignore that.

We also needed to update the `:update-available?` subscription to better
take into account the first run (IE: *no* update seen), and the case
where we have a new version, but don't yet know what version we've seen.