Refactor and simplify limited-select widget
Refs #14

Significantly improves render times, dropping React from ~519ms down to
~400ms, and views from ~457ms down to a much more palatable ~150ms.

The key here was reducing a lot of cruft and overhead from insisting on
using the reagent-forms, which it really was obnoxiously unnecessary
here. Instead, we just replicate some of the plumbing using normal
re-frame components, which has the added bonus of simplifying a lot of
code.

We ought to be able to further optimize this by virtualizing the list of
displayed elements, now that we're not shackled to reagent-forms.