Virtualize scrollable limited-select widgets
Refs #14

This drops View rendering further down from ~150ms to just ~50ms,
with React crushed from ~400ms down to ~120ms.

So this is pretty great, and we didn't have to resort to dropdowns!
There are still some visual niggles that it'd be nice to fix (in
particular, the bottoms of lists seem to get cut off a bit) but this
seems like a great technique to use here. We probably won't even need to
collapse the features (though we might want to in a future change for a
more glanceable visual).