Optimize Builder's rendering of features with many options (#89)
* Refactor and simplify limited-select widget

Refs #14

Significantly improves render times, dropping React from ~519ms down to
~400ms, and views from ~457ms down to a much more palatable ~150ms.

The key here was reducing a lot of cruft and overhead from insisting on
using the reagent-forms, which it really was obnoxiously unnecessary
here. Instead, we just replicate some of the plumbing using normal
re-frame components, which has the added bonus of simplifying a lot of
code.

We ought to be able to further optimize this by virtualizing the list of
displayed elements, now that we're not shackled to reagent-forms.

* Virtualize scrollable limited-select widgets

Refs #14

This drops View rendering further down from ~150ms to just ~50ms,
with React crushed from ~400ms down to ~120ms.

So this is pretty great, and we didn't have to resort to dropdowns!
There are still some visual niggles that it'd be nice to fix (in
particular, the bottoms of lists seem to get cut off a bit) but this
seems like a great technique to use here. We probably won't even need to
collapse the features (though we might want to in a future change for a
more glanceable visual).

* Simplify virtual-list and fix padding/margin issues

Refs #14

* Clean up docs, add overscan-row-count option (with default 10)

Refs #14

* Remove old, unused reagent-forms limited-select widget

We could always refactor our new one into a separate namespace if needed