Add some placeholders and a link to the Campaign UI ticket
It's about time for this branch to get merged; there's a ton of
refactoring in it already; all the remaining features can be done
in one or more seperate branches.

The only thing we really need to do before merging is allow players
to detach the campaign from their character.