Update dependencies
I'd been using a somewhat dated version of leiningen, and recently
updated it when working on an unrelated project. I came back here and
discovered my dev environment was messed up, so started exploring.

Previously, I'd avoided updating updating `cemerick/piggieback` ->
`cider/piggieback` because doing so caused other issues in my dev
environment, but I suspect those were caused by the older version of
lein that I was using. Now that I'm on the latest version I can finally
update, and everything seems to be working well for me again!

Another artifact of updating is that the lein-npm plugin is showing
how unmaintained it is, with warnings about hooks being deprecated.
Luckily, we can just use `:npm-deps` on the `test` build to install
karma and friends; that's all we were using `lein-npm` for anyway.