Query cached sheets first (#107)
* Support multiple results from query-sheets channels

Refs #105

This also returns cached results *first* without waiting on any results
from the original provider. This does mean, however, that we need to
clear the cached sheets on signout. (We probably should have been
doing this anyway.)

* Clear loaded and cached sheets on disconnect of provider

We should have been doing this before, but it's especially important
due to the changes in #105 to list sheets from cache immediately.