Dynamic "Temporary Effects" system (#121)
* Begin work on compiler and subscription support for "effects"

Refs #113

We need a way to add them... and probably a way to visualize the
changes they make.

* Fix effect inflation + subs; add events for adding/removing

Refs #113

* Add initial "buffed" indications + an "Affected by" row

Includes support for doubling speed with an effect. Items in the
Affected by row can be tapped/clicked to view the full information,
including bullets from a special `:effects` key. This overlay
should probably support ending the effect.

* Support ending an effect from its info overlay

* Create UI for adding an effect

Refs #113

We're now feature complete, but one last nicety would be support
for automatically adding an effect from a limited-use or self-cast
spell.

* Support auto-adding effect from a limited-use!

Refs #113

* Don't show :feature-only? effects in the search window

* Support triggering effects from the single-usage box

Refs #113