Don't force :desc to always be callable, if it's a const val
This is not *great* but fine for stability right now. Basically, :desc
for some feature options (I think) are getting turned into fns when they
oughtn't be, but also aren't then getting run to transform back into a
value.

At some point we could really benefit from some major refactoring to
simplify this core system because it's quite complicated and it feels
like there ought to be a better way to do some of these things...