Initial refactor from cljs-css-modules -> spade
Our old macros were basically the same ideas behind spade, so this was
mostly a one-to-one translation. There ought to be some opportunities
for improvement here as well, though. At the very least, we could move
view-specific styles into the view ns, instead of keeping everything in
a single ns.