Refactor to begin using spade for styling (#133)
* Initial refactor from cljs-css-modules -> spade

Our old macros were basically the same ideas behind spade, so this was
mostly a one-to-one translation. There ought to be some opportunities
for improvement here as well, though. At the very least, we could move
view-specific styles into the view ns, instead of keeping everything in
a single ns.

* Move abilities-related styles into the namespace

* Move header styles into the header view ns

* Move all overlay styles into their own ns (for now)

* Update style references

* Move at-media specs to their own ns + other cleanup