5e: Fix sorcery point level scaling + add custom :restore-desc
Fixes #139

Any limited-use whose uses description can't be automatically computed,
such as this one, can now provide an explicit string.

Also, it's good to remember that we don't necessarily need functions
for level scaling stuff, since level will be provided as part of the
state when the feature is provided!