5e Experiment: fix nested button activation by fake nesting
Currently, when the "Cast" button is pressed on the header of a spell,
because that button is nested within the button header for the spell
both elements get the :active state, which looks a bit weird.

This change solves that by instead rendering the Cast button as an
absolutely-positioned sibling, while also rendering the same button
as a nested child, but invisible, to keep the spacing as expected.

This might be a terrible idea.