Simplify builder navigation UX
Instead of the named "arrows" that sort of float somewhere in the middle
of the page, we now have *actual* arrows that stick to the top, easily
accessible. Section headers are still available for quick nav, but we
may consider simplified names or even icons to save space on smaller
screens.