Add initial "availability group" concept; show "other" items in group
See #138

If a feature has an `:availability-attr` that is a vector, the first
keyword in that vector becomes its "availability group."

A feature implicitly "has" an availability group if its first two
options are part of the same availability group. Alternately, a feature
could manually specify which availability groups it contributes to
by specifying the `:availability-groups` key. This is useful for things
like the "pick any 2" for a custom background, which allows both tool
proficiencies and languages.